requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
qrcode==8.2
Pillow==11.3.0
python-dotenv==1.0.0
Werkzeug==3.1.3
openpyxl==3.1.2
pytest==7.4.3
twilio==9.3.3
gunicorn==23.0.0
